src/server/requirements.txt
anyio==4.4.0
asgiref==3.8.1
boto3==1.34.147
botocore==1.34.147
click==8.1.7
fastapi==0.95.2
h11==0.14.0
idna==3.7
jmespath==1.0.1
localstack-client==2.5
mypy-boto3-athena==1.34.130
mypy-boto3-glue==1.34.143
mypy-boto3-s3==1.34.138
numpy==2.0.1
pandas==2.2.2
pip==24.0
pydantic==1.10.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.5
pytz==2024.1
s3transfer==0.10.2
setuptools==65.5.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.15.0
httpx==0.27.0
python-jose==3.3.0